File Merging
Merged Zrupi UI with main file.

Fixed errors with DBHandler addEvent call.
